proc: define PRIO_OFFSET, MIN_PRIO, MAX_PRIO as int, fix assertions
TASK: RTOS-1434

+++ syscalls.c
@@ -320,7 +320,7 @@ int syscalls_beginthreadex(u8 *ustack)
 		return -EFAULT;
 	}
 
-	if (priority < (int)MIN_PRIO || priority > (int)MAX_PRIO) {
+	if (priority < MIN_PRIO || priority > MAX_PRIO) {
 		return -EINVAL;
 	}
